+++ waspc/data/packages/deploy/src/providers/railway/commands/setup/SetupCmdOptions.ts
@@ -8,4 +8,5 @@ export interface SetupCmdOptions extends CommonCmdOptions, SecretsOptions {
   existingProjectId?: RailwayProjectId;
   workspace?: string;
   dbImage: string;
+  dbVolumeMountPath: string;
 }