+++ waspc/src/Wasp/Db/Postgres.hs
@@ -23,10 +23,8 @@ postgresMaxDbNameLength = 63
 -- is released. We aim to occasionally update this version in Wasp releases.
 -- If you bump the Postgres version here, also update the default database images in
 -- `waspc/data/packages/deploy/src/providers/fly/index.ts` and
--- `waspc/data/packages/deploy/src/providers/railway/index.ts`, and check if
--- `dockerVolumeMountPath` here and `databaseVolumeMountPath` in
--- `waspc/data/packages/deploy/src/providers/railway/railwayService/database.ts`
--- are still correct.
+-- `waspc/data/packages/deploy/src/providers/railway/index.ts`, and check that
+-- their default volume mount paths are still correct.
 defaultPostgresDockerImageSpec :: (DockerImageName, DockerVolumeMountPath)
 defaultPostgresDockerImageSpec = ("postgres:18", dockerVolumeMountPath)
   where